Move storage fixtures to fixtures.py
